Explicit types and added comment

+++ packages/contracts/src/test/dao/GasConsumerHelper.sol
@@ -2,11 +2,12 @@
 
 pragma solidity 0.8.17;
 
+/// @notice This contract is used for testing to consume gas.
 contract GasConsumer {
     mapping(uint256 => uint256) public store;
 
     function consumeGas(uint256 count) external {
-        for (uint i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             store[i] = 1;
         }
     }
